Improved log file header
I changed the redirection method from overwrite to append (to stop the accidental erasure of previous event logs on the script being re-run. There were unnecessary characters to specify a return carriage at the end but those were not needed as a standard echo achieved the same, so I removed those.